dndBattleDice v1.0.0 Initial Commit
With maintainability in mind, I have set up the basis of the bot logic, storage and presentation layers with the initial commit.

The main dice roll functionality is scheduled for the next commit, wherein users can create profiles with the bot to store their dnd player character data within the backend of the application and my roll commands will pull from this data layer.

For now, I have developed commands of different access permission levels to be used by guild members that allow adjustment of the bot configuration within the guild where the change request originates, and this data is stored in the backend and pulled from to determine how the bot functions within each guild according to preferences and enabled features.

I have also included a help command and introductory message referencing the help section, so that guild members can quickly learn how to best utilise the bot.